[#15] refactor: Photo의 로직을 usePhoto 커스텀훅으로 분리
